+++ wp-rest-api-v2-custom-post-fields/plugin.php
@@ -1,11 +1,11 @@
 <?php
 
 /**
- * Plugin Name: WP REST API V2 Custom Post Types
- * Description: Adds Advanced Custom Post Types to WP REST API V2 JSON output.
- * Version: 0.1
+ * Plugin Name: WP REST API V2 Custom Post Fields
+ * Description: Adds Advanced Custom Post Fields to WP REST API V2 JSON output.
+ * Version: 0.2
  * Author: Deyan Vatsov
- * Plugin URI: https://github.com/Vatsov/wp-rest-api-v2-custom-post-types/
+ * Plugin URI: https://github.com/Vatsov/wp-rest-api-v2-custom-post-fields/
  */
 
 if ( !function_exists( 'is_plugin_active' ) ) {
@@ -19,7 +19,7 @@ if ( is_plugin_active('rest-api/plugin.php') ) {
 class CustomPostDataPlugin {
 	public function __construct() {
 		// Add Meta Fields to Posts
-		add_action('rest_api_init',		array( $this, 'add_custom_data' ) );
+		add_action('rest_api_init', 	array( $this, 'add_custom_data' ) );
 		add_filter('rest_query_vars',	array( $this, 'allow_meta_query' ) );
 		add_filter('rest_post_query',	array( $this, 'allowed_meta_keys' ) );
 	}
@@ -59,7 +59,7 @@ class CustomPostDataPlugin {
 		return update_post_meta($post->ID, $field_name, strip_tags($value));
 	}
 
-	//Example: /wp-json/wp/v2/posts?filter[meta_key]=slideshow&filter[meta_value]=true&filter[cast]=boolean
+	//Example: /wp-json/wp/v2/posts?filter[meta_key]=slideshow&filter[meta_value]=true
 	function allow_meta_query($valid_vars) {
 		$valid_vars = array_merge( $valid_vars, array( 'meta_key', 'meta_value' ) );
 		return $valid_vars;
@@ -67,19 +67,20 @@ class CustomPostDataPlugin {
 
 	function allowed_meta_keys($args) {
 		// ToDo: to be configurable from admin panel
+		// set the allowed meta keys or asterisk for grant access
 		$allowed_meta_keys = array( 'slideshow' );
 
-		if ( $args['cast'] === "boolean" ) {
-			if ( $args['meta_value'] === "true" ) {
-				$args['meta_value'] = true;
-			}
 
-			if ( $args['meta_value'] === "false" ) {
-				$args['meta_value'] = false;
-			}
+		if ( isset($args['meta_value']) && $args['meta_value'] === "true" ) {
+			$args['meta_value'] = true;
 		}
 
-		if ( !in_array( $args['meta_key'], $allowed_meta_keys ) ) {
+		if ( isset($args['meta_value']) && $args['meta_value'] === "false" ) {
+			$args['meta_value'] = false;
+		}
+
+
+		if ( $allowed_meta_keys[0] !== '*' && isset($args['meta_key']) && !in_array( $args['meta_key'], $allowed_meta_keys ) ) {
 			unset( $args['meta_key'] );
 			unset( $args['meta_value'] );
 		}