+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # Changelog
+* (29 April 2016). Allow filter by meta key/value.
 * (10 April 2016). Initial Release.